table-browser: move all enhanced analysisi modules to enhancedanalysis subdir
