+++ src/scripts/modules/table-browser/react/components/EnhancedAnalysis/EnhancedAnalysisRunControl.jsx
@@ -1,5 +1,5 @@
 import React, {PropTypes} from 'react';
-import immutableMixin from '../../../../react/mixins/ImmutableRendererMixin';
+import immutableMixin from '../../../../../react/mixins/ImmutableRendererMixin';
 import {Loader} from 'kbc-react-components';
 import {Button} from 'react-bootstrap';
 import {Link} from 'react-router';
